requirements.txt
streamlit
ell-ai
openai
python-dotenv
